Another-little revolution, a summary of changes: 1) Reorganized directory structure using a molsim directory 2) Visualizer implemented with the concept of a list of "renderers" 3) some forcers, integrators etc.

+++ chemlab/core/molecule.py
@@ -158,7 +158,7 @@ class Atom:
             Atom._curid += 1
         
         self.type = type
-        self.coords = np.array(coords)
+        self.coords = np.array(coords, dtype=np.float32)
 
         self.atno = symbols.symbol_list.index(type) + 1
